move dataset classes to datasets.py
